pyproject.toml
[tool.poetry]
name = "rentapp"
version = "0.1.0"
description = ""
authors = ["umerfayyaz97 <115028799+umerfayyaz97@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.19"
fastapi = "^0.110.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
